Once again show campsites as we're loading the data
This is because we're now putting the park name in the campsite object.
This means that a campsite is now updated automatically which makes
the user display much less chaotic. On balance it's much better for
a user to get a sense that the data is loading than not. So, we're
displaying it again.